Changed all score values to Integer to keep things consistent - forgot why i made them floats in the first place. working fine now after fixing a core data bug

+++ Jest Notes/Joke.h
@@ -13,7 +13,7 @@
 @interface Joke : NSObject
 
 @property (nonatomic, strong) NSString* name;
-@property float score;
+@property int score;
 @property int length;
 @property (nonatomic, strong) NSDate *writeDate;
 @property (nonatomic, strong) NSString* bodyText;
